Add bones polygons as internal nodes
When copying/duplicating in the editor, nodes children are also duplicated and
this is a problem because EVERYTHING (`_ready`, `_init`, `_enter_tree`) is called BEFORE
the variables that hold this nodes are setted (eg `_bone_shape`).

So _bone_shape will be null but the node exist and I don't have how to link them
without guessing their order (which is not safe).

"nodes children are also duplicated" except when they are internal, which
make totally protect us from this problem (I'm open for better solution).